+++ app/vite.config.ts
@@ -5,6 +5,9 @@ import tailwindcss from '@tailwindcss/vite';
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [vue(), tailwindcss()],
+  build: {
+    assetsDir: 'app-assets',
+  },
   server: {
     port: 5173,
     proxy: {